Convert tickets row generators to hooks
