--- assets/src/domain/eventEditor/ui/tickets/hooks/useTicketRegistrationCount.ts
@@ -1,8 +1,0 @@
-import { EntityId } from '@appServices/apollo/types';
-
-const useTicketRegistrationCount = (ticketId: EntityId): number => {
-	// @todo get registration count
-	return ticketId.length;
-};
-
-export default useTicketRegistrationCount;
