+++ assets/src/domain/eventEditor/ui/tickets/ticketsList/tableView/TableView.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { __ } from '@wordpress/i18n';
 
 import { EntityTable } from '@appLayout/entityList';
-import headerRowGenerator from './headerRowGenerator';
-import bodyRowGenerator from './bodyRowGenerator';
+import useHeaderRowGenerator from './useHeaderRowGenerator';
+import useBodyRowGenerator from './useBodyRowGenerator';
 import { useTicketsListContext } from '@edtrServices/context/EntityListContext';
 
 /**
@@ -11,6 +11,9 @@ import { useTicketsListContext } from '@edtrServices/context/EntityListContext';
  */
 const TableView: React.FC = () => {
 	const { filterState, filteredEntities } = useTicketsListContext();
+
+	const bodyRowGenerator = useBodyRowGenerator();
+	const headerRowGenerator = useHeaderRowGenerator();
 
 	return (
 		<EntityTable
